wow that really shit-blasted on CPU, thanks electron

+++ test-grabity.mjs
@@ -4,8 +4,10 @@ import fs from 'fs';
 import nss from 'nightmare-screenshot-selector';
 Nightmare.action('screenshotSelector', nss);
 let busy = 0;
+const nightmare = Nightmare({
+  waitTimeout: 60000
+});
 const test = async () => {
-  const nightmare = Nightmare({ show: false });
   nightmare.goto('https://twitter.com/dril/status/1331742968709926912')
     .wait('article[data-focusable=true][tabindex="0"] div div div div div div span')
     .screenshotSelector('article[data-focusable=true][tabindex="0"]')
@@ -23,14 +25,16 @@ const test = async () => {
 
 const fire = (url, msgid, idx, res2) => {
   return new Promise((res, rej) => {
+    if(fs.existsSync(`./cache/${msgid}-${idx}.png`)){
+      return res({msgid, idx, url: `cache/${msgid}-${0}.png`, msg: 'shotted', type: 'twitter'});
+    }
     if(busy === 0){
       busy = 1;
       const nightmare = Nightmare({ show: false });
       nightmare.goto(url)
         .wait('article[data-focusable=true][tabindex="0"] div div div div div div span')
         .screenshotSelector('article[data-focusable=true][tabindex="0"]')
-        .end()
-        .then(snap=>{
+        .then(snap => {
           fs.writeFileSync(`./cache/${msgid}-${idx}.png`, snap);
           busy = 0;
           const goop = {
@@ -43,8 +47,7 @@ const fire = (url, msgid, idx, res2) => {
           res2 ? res2(goop) : res(goop);
         })
         .catch(error => {
-          console.log('twit fetch failed', error);
-          nightmare.end();
+          console.log('twit fetch failed', url, msgid, error);
           busy = 0;
           const goop = {
             msgid,
